moved contextmanager data base to conftest.py
